separated filterbtns comp
