+++ src/Containers/SubscriptionListContainer/SubscriptionListContainer.tsx
@@ -1,23 +1,18 @@
 import React, { useState } from "react";
 import { Button } from "@skbkontur/react-ui/components/Button";
 import AddIcon from "@skbkontur/react-icons/Add";
-import { Filter } from "@skbkontur/react-icons";
 import { Subscription } from "../../Domain/Subscription";
 import { createSchedule, WholeWeek } from "../../Domain/Schedule";
 import { Contact } from "../../Domain/Contact";
 import SubscriptionEditModal from "../../Components/SubscriptionEditModal/SubscriptionEditModal";
 import CreateSubscriptionModal from "../../Components/CreateSubscriptionModal/CreateSubscriptionModal";
 import type { SubscriptionInfo } from "../../Components/SubscriptionEditor/SubscriptionEditor";
-import TagDropdownSelect from "../../Components/TagDropdownSelect/TagDropdownSelect";
 import { SubscriptionList } from "../../Components/SubscriptionList/SubscriptionList";
-import { filterSubscriptions } from "../../Domain/FilterSubscriptions";
 import { AddSubscriptionMessage } from "../../Components/AddSubscribtionMessage/AddSubscribtionMessage";
 import { ModalType } from "../../Domain/Global";
 import { ConfigState } from "../../store/selectors";
 import { useAppSelector } from "../../store/hooks";
-import { Checkbox, DropdownMenu, MenuItem, ThemeContext, ThemeFactory } from "@skbkontur/react-ui";
-import ContactTypeIcon from "../../Components/ContactTypeIcon/ContactTypeIcon";
-import { ArrowChevronDown } from "@skbkontur/react-icons";
+import { FilterSubscriptionButtons } from "./Components/FilterSubscriptionButtons";
 import classNames from "classnames/bind";
 
 import styles from "./SubscriptionListContainer.less";
@@ -70,19 +65,14 @@ export const SubscriptionListContainer: React.FC<Props> = (props) => {
 
     const [newSubscription, setNewSubscription] = useState<SubscriptionInfo | null>(null);
     const [subscriptionToEdit, setSubscriptionToEdit] = useState<Subscription | null>(null);
-    const [filterTags, setFilterTags] = useState<string[]>([]);
-    const [filterContactIDs, setFilterContactIDs] = useState<string[]>([]);
 
-    const { filteredSubscriptions, availableTags, availableContactIDs } = filterSubscriptions(
+    const { FilterButtons, filteredSubscriptions } = FilterSubscriptionButtons({
+        contacts,
         subscriptions,
-        filterTags,
-        filterContactIDs
-    );
+    });
+
     const { config } = useAppSelector(ConfigState);
 
-    const availableContacts = contacts.filter((contact) =>
-        availableContactIDs?.includes(contact.id)
-    );
     const isPlottingDefaultOn = !!config?.featureFlags.isPlottingDefaultOn;
 
     const handleCloseModal = (modal: ModalType) => {
@@ -95,10 +85,6 @@ export const SubscriptionListContainer: React.FC<Props> = (props) => {
                 setSubscriptionToEdit(null);
                 break;
         }
-    };
-
-    const handleFilterTagsChange = (tags: string[]): void => {
-        setFilterTags(tags);
     };
 
     const handleEditSubscription = (subscription: Subscription): void => {
@@ -180,19 +166,6 @@ export const SubscriptionListContainer: React.FC<Props> = (props) => {
         }
     };
 
-    const renderFilterByContactCaption = ({ openMenu }: { openMenu: () => void }) => {
-        return (
-            <Button
-                width={180}
-                icon={filterContactIDs.length ? <Filter /> : <ArrowChevronDown />}
-                use="default"
-                onClick={openMenu}
-            >
-                Filter by contact
-            </Button>
-        );
-    };
-
     return (
         <>
             {subscriptions.length > 0 ? (
@@ -208,56 +181,7 @@ export const SubscriptionListContainer: React.FC<Props> = (props) => {
                             >
                                 Add subscription
                             </Button>
-                            <DropdownMenu
-                                menuMaxHeight={300}
-                                caption={renderFilterByContactCaption}
-                            >
-                                {availableContacts.map((contact) => (
-                                    <ThemeContext.Provider
-                                        key={contact.id}
-                                        value={ThemeFactory.create({
-                                            menuItemHoverBg: "initial",
-                                        })}
-                                    >
-                                        <MenuItem>
-                                            <Checkbox
-                                                checked={filterContactIDs.includes(contact.id)}
-                                                style={{
-                                                    alignItems: "center",
-                                                }}
-                                                onValueChange={() => {
-                                                    const contactIndex = filterContactIDs.indexOf(
-                                                        contact.id
-                                                    );
-                                                    if (contactIndex === -1) {
-                                                        setFilterContactIDs((prev) => [
-                                                            ...prev,
-                                                            contact.id,
-                                                        ]);
-                                                    } else {
-                                                        setFilterContactIDs((prev) =>
-                                                            prev.filter((id) => {
-                                                                return id !== contact.id;
-                                                            })
-                                                        );
-                                                    }
-                                                }}
-                                            >
-                                                <ContactTypeIcon type={contact.type} />
-                                                &nbsp;
-                                                {contact.value}
-                                            </Checkbox>
-                                        </MenuItem>
-                                    </ThemeContext.Provider>
-                                ))}
-                            </DropdownMenu>
-                            <TagDropdownSelect
-                                width={180}
-                                value={filterTags}
-                                availableTags={availableTags}
-                                onChange={handleFilterTagsChange}
-                                placeholder="Filter by tag"
-                            />
+                            {FilterButtons}
                         </div>
                     </div>
                     <SubscriptionList
